Update required status checks for R-guide repo
The R-guide repo deploys to GitHub Pages from the main branch. GitHub
Pages does not have preview deployments, so it is not possible to deploy
before merging to main. The required status checks (on PRs for example)
should only require the build status check. The deployment will commence
after merging to main.