Simplify IAM policy output
The `github-actions-oidc-provisioning` IAM policy was previously output
so it could be used in another stack (the `terraform-cloud-oidc` stack).
Although that stack is no longer in use, the output will be maintained
in case it is useful to other stacks in the future.

The `attachment_count` attribute of `aws_iam_policy` may change outside
of this stack, and may therefore induce drift in the output. This commit
will omit the `attachment_count` from the output to avoid this drift.