Run OpenTofu in GitHub Actions with Digger
This commit will configure Digger to run OpenTofu in GitHub Actions.

Digger requires the workflow file to be named `digger_workflow.yml`.

Note the `permissions:` configuration, which sets up permissions for the
`GITHUB_TOKEN`. As the GitHub docs explain, "If you specify the access
for any of these permissions, all of those that are not specified are
set to `none`."

https://docs.digger.dev/readme/introduction
https://docs.digger.dev/ce/getting-started/with-opentofu
https://docs.github.com/en/actions/reference/workflows-and-actions/workflow-syntax