Allow Spacelift stacks to read AWS stack state
This commit will add an `aws_iam_role_policy_attachment` resource to the
IAM role that will be used in the AWS integration for Spacelift stacks.
This attachment allows the AWS IAM role used for the Spacelift space to
access state from AWS stacks. This access is needed for provisioning
`spacelift_aws_integration` resources.