Configure Spacelift AWS integration credentials
This commit will configure stacks to use the AWS IAM roles provided by
the Spacelift AWS integrations in each Spacelift space. Configurations
will use either dynamic OIDC credentials or static credentials depending
on the value of an input variable specifying an OIDC role ARN.

Roles must first be provisioned by applying the OpenTofu configurations
for spacelift-oidc. This presents a chicken-and-egg problem - the
configurations will provision roles for all stacks to use, but the roles
will not exist until the configurations are applied - so how do these
configurations themselves get credentials? One solution is to apply the
configurations locally or with static AWS access key credentials the
first time to provision the roles that will be used in later applies.

After roles are provisioned, the Spacelift provider configurations for
the given account must then be applied to provision AWS integrations,
contexts, and environment variables for the Spacelift account. The
Spacelift OIDC role ARNs will be used to provision the integrations, and
environment variables with the role ARNs will be set in each context.
The role ARNs can be passed to the Spacelift provider configurations
with `terraform_remote_state` data sources, Spacelift stack
dependencies, or another preferred method. The Spacelift provider
configurations present another chicken-and-egg problem - if role ARNs
for Spacelift integrations are read from remote state, then reading
remote state requires credentials from a role ARN. Again here, one
solution is to apply the configurations locally or with static
credentials the first time.