Add Spacelift credential environment variables
The default credentials for Spacelift administrative stacks only allow
them to administrate the space they are in. Therefore, it can be helpful
to either put administrative stacks in the `root` space or to generate
an admin API key for the administrative stack so it can access all
resources in the Spacelift account. Otherwise, its access may be limited
by the space that the stack is in.

https://docs.spacelift.io/concepts/stack/stack-settings#administrative