Move GitHub Actions OIDC policy to Spacelift OIDC
This commit will move the GitHub Actions OIDC policy (again) from the
github-actions-oidc stack to the spacelift-oidc stack. This is mostly to
avoid chicken-and-egg problems when setting up the repo. The roles used
for credentials are provisioned by the spacelift-oidc stack, so it is
simplest to also attach policies there. Otherwise, the spacelift-oidc
stack would have to reference a policy from the github-actions-oidc
stack that might not exist yet.