Build custom Docker image for Spacelift
The 1Password `op` CLI is required for the OpenTofu provider, and the
GitHub admin stack uses the 1Password provider, so the GitHub stack
needs `op`. It would be simplest to install `op` in a `before_init`
stack hook, but Spacelift runner images use a non-root user `spacelift`
that cannot install apk repositories, so a custom image build is needed.

This commit will add a Dockerfile, a GitHub Actions job to build the
image specified in the Dockerfile, and an AWS ECR Public repository to
use for the image in addition to GitHub Container Registry (GHCR).

ECR Public resources only support us-east-1.

The 1Password CLI (op) does not support arm64 (aarch64). Attempting to
install op on linux/arm64 with apk may result in the following warning:

```text
WARNING: updating and opening
https://downloads.1password.com/linux/alpinelinux/stable/:
No such file or directory
```

- https://developer.1password.com/docs/cli/get-started/
- https://developer.1password.com/docs/terraform/?workflow-type=1password-cli
- https://docs.spacelift.io/concepts/stack/stack-settings.html#customizing-workflow