Add this repo to GitHub stack
This commit will add br3ndonland/infrastructure-live to the GitHub stack
and will import existing repo resources. Note that repos must be public
to use repo rulesets. This repo was initially private while it was being
set up to ensure it was configured securely first, and also because the
repo was not necessarily going to be used or maintained. Now that it is
configured and being maintained, it's fine to make it public.